Add Copy link to Post Controls for Post List
This should be reviewed from a design standpoint, as adding a new
control would relegate the Trash control (arguably more important and
used than Copy) inside the More sub-menu.
At the same time, Copy as the last control (stowed in the More sub-menu)
seems to me quite out of place.